Rename MoviewInfoForm -> MovieInfoContainer in React demo
