+++ apps/demos/Demos/Scheduler/Templates/React/MovieInfoContainer.tsx
